Re-run test file if it has changed
The test cases inside a test file might change even if the dependencies haven't changed. In that scenario the plugin should consider run all test cases. As we can't calculate the hash of a test method or know if something outside the test method but inside the test file has changed.

Closes #8